[ANDROSDK-1131] Refactor to TrackedEntityInstanceService
