+++ core/src/main/java/org/hisp/dhis/android/core/relationship/RelationshipModule.java
@@ -31,6 +31,4 @@ public interface RelationshipModule {
     RelationshipTypeCollectionRepository relationshipTypes();
 
     RelationshipCollectionRepository relationships();
-
-    RelationshipDomain domain();
 }
