+++ core/src/main/java/org/hisp/dhis/android/core/relationship/internal/RelationshipModuleImpl.java
@@ -28,7 +28,6 @@
 package org.hisp.dhis.android.core.relationship.internal;
 
 import org.hisp.dhis.android.core.relationship.RelationshipCollectionRepository;
-import org.hisp.dhis.android.core.relationship.RelationshipDomain;
 import org.hisp.dhis.android.core.relationship.RelationshipModule;
 import org.hisp.dhis.android.core.relationship.RelationshipTypeCollectionRepository;
 
@@ -43,15 +42,11 @@ public final class RelationshipModuleImpl implements RelationshipModule {
 
     private final RelationshipCollectionRepository relationships;
 
-    private final RelationshipDomain domain;
-
     @Inject
     RelationshipModuleImpl(RelationshipTypeCollectionRepository relationshipTypeRepository,
-                           RelationshipCollectionRepository relationshipRepository,
-                           RelationshipDomain relationshipDomain) {
+                           RelationshipCollectionRepository relationshipRepository) {
         this.relationshipTypes = relationshipTypeRepository;
         this.relationships = relationshipRepository;
-        this.domain = relationshipDomain;
     }
 
     @Override
@@ -62,10 +57,5 @@ public final class RelationshipModuleImpl implements RelationshipModule {
     @Override
     public RelationshipCollectionRepository relationships() {
         return relationships;
-    }
-
-    @Override
-    public RelationshipDomain domain() {
-        return domain;
     }
 }
