+++ core/src/main/java/org/hisp/dhis/android/core/trackedentity/TrackedEntityModule.java
@@ -45,4 +45,6 @@ public interface TrackedEntityModule {
     TrackedEntityAttributeReservedValueManager reservedValueManager();
 
     TrackedEntityInstanceDownloader trackedEntityInstanceDownloader();
+
+    TrackedEntityInstanceService trackedEntityInstanceService();
 }
